Add facebook accountKit module and refactoring packages
