+++ api/build.gradle
@@ -34,11 +34,15 @@ dependencies {
             'org.springframework.boot:spring-boot-starter-validation:1.5.11.BUILD-SNAPSHOT',
             'org.springframework.boot:spring-boot-starter-web:1.5.11.BUILD-SNAPSHOT',
             'org.springframework.boot:spring-boot-starter-data-jpa:1.5.11.BUILD-SNAPSHOT',
-            'org.springframework.boot:spring-boot-starter-freemarker:1.5.11.BUILD-SNAPSHOT'
+            'org.springframework.boot:spring-boot-starter-freemarker:1.5.11.BUILD-SNAPSHOT',
+            'org.springframework.cloud:spring-cloud-starter-hystrix:1.3.4.RELEASE'
 
     compile 'org.springframework.boot:spring-boot-starter-security:1.5.10.RELEASE',
             'org.springframework.security.oauth:spring-security-oauth2:2.3.0.RELEASE',
             'org.springframework.security:spring-security-jwt:1.0.9.RELEASE'
+
+    compile 'org.springframework.cloud:spring-cloud-starter-openfeign:1.4.4.RELEASE'
+    compile 'io.github.openfeign:feign-httpclient:9.5.1'
 
     compile 'io.jsonwebtoken:jjwt:0.7.0'
