+++ api/src/main/java/me/nexters/liliput/api/ApiApplication.java
@@ -1,19 +1,22 @@
 package me.nexters.liliput.api;
 
-import me.nexters.liliput.api.common.util.ShortUrlBuilder;
+import me.nexters.liliput.api.api.ApiConfig;
+import me.nexters.liliput.api.application.ApplicationConfig;
+import me.nexters.liliput.api.domain.DomainConfig;
+import me.nexters.liliput.api.infrastructure.InfrastructureConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
+@Import({
+            ApiConfig.class,
+            ApplicationConfig.class,
+            DomainConfig.class,
+            InfrastructureConfig.class
+        })
 public class ApiApplication {
     public static void main(String[] args) {
         SpringApplication.run(ApiApplication.class, args);
-    }
-
-    @Bean
-    public ShortUrlBuilder shortUrlBuilder() {
-        return new ShortUrlBuilder();
     }
 }
