+++ api/src/main/java/me/nexters/liliput/api/api/ApiConfig.java
@@ -1,0 +1,15 @@
+package me.nexters.liliput.api.api;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+import org.springframework.cloud.netflix.feign.EnableFeignClients;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+@EnableFeignClients
+@EnableCircuitBreaker
+@ComponentScan
+public class ApiConfig {
+}
