+++ api/src/main/java/me/nexters/liliput/api/api/feign/facebook/dto/DeleteAccountResponseResponse.java
@@ -1,0 +1,10 @@
+package me.nexters.liliput.api.api.feign.facebook.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+public class DeleteAccountResponseResponse {
+    @JsonProperty("success")
+    private boolean removed;
+}
