+++ api/src/main/java/me/nexters/liliput/api/application/ApplicationConfig.java
@@ -1,0 +1,17 @@
+package me.nexters.liliput.api.application;
+
+import me.nexters.liliput.api.application.common.util.ShortUrlBuilder;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+@ComponentScan
+public class ApplicationConfig {
+    @Bean
+    public ShortUrlBuilder shortUrlBuilder() {
+        return new ShortUrlBuilder();
+    }
+}
