+++ api/src/main/java/me/nexters/liliput/api/application/common/constant/Constants.java
@@ -1,0 +1,4 @@
+package me.nexters.liliput.api.application.common.constant;
+
+public class Constants {
+}
