--- api/src/main/java/me/nexters/liliput/api/common/constant/Constants.java
@@ -1,4 +1,0 @@
-package me.nexters.liliput.api.common.constant;
-
-public class Constants {
-}
