+++ api/src/main/java/me/nexters/liliput/api/domain/DomainConfig.java
@@ -1,0 +1,11 @@
+package me.nexters.liliput.api.domain;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+@ComponentScan
+public class DomainConfig {
+}
