+++ api/src/main/java/me/nexters/liliput/api/domain/dto/ShortUrlModel.java
@@ -1,0 +1,9 @@
+package me.nexters.liliput.api.domain.dto;
+
+import lombok.Data;
+
+@Data
+public class ShortUrlModel {
+    private String path;
+    private String webUrl;
+}
