--- api/src/main/java/me/nexters/liliput/api/domain/route/repository/RouteLog.java
@@ -1,27 +1,0 @@
-package me.nexters.liliput.api.domain.route.repository;
-
-import lombok.Data;
-import me.nexters.liliput.api.domain.shared.BaseEntity;
-import me.nexters.liliput.api.domain.shorten.repository.ShortUrl;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import java.io.Serializable;
-
-@Data
-@Entity
-@Table(name = "ROUTE_LOG")
-public class RouteLog extends BaseEntity implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    @ManyToOne
-    private ShortUrl shortUrl;
-    private String referer;
-    private String userAgent;
-    private String remoteAddress;
-}
