--- api/src/main/java/me/nexters/liliput/api/domain/route/repository/RouteLogRepository.java
@@ -1,6 +1,0 @@
-package me.nexters.liliput.api.domain.route.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RouteLogRepository extends JpaRepository<RouteLog, Long> {
-}
