--- api/src/main/java/me/nexters/liliput/api/domain/route/service/RouteLogService.java
@@ -1,5 +1,0 @@
-package me.nexters.liliput.api.domain.route.service;
-
-public interface RouteLogService {
-    void logging(String path, String referer, String userAgent, String remoteAddress);
-}
