+++ api/src/main/java/me/nexters/liliput/api/domain/service/RouteLogService.java
@@ -1,0 +1,5 @@
+package me.nexters.liliput.api.domain.service;
+
+public interface RouteLogService {
+    void logging(String path, String referer, String userAgent, String remoteAddress);
+}
