--- api/src/main/java/me/nexters/liliput/api/domain/shorten/service/ShortUrlModel.java
@@ -1,9 +1,0 @@
-package me.nexters.liliput.api.domain.shorten.service;
-
-import lombok.Data;
-
-@Data
-public class ShortUrlModel {
-    private String path;
-    private String webUrl;
-}
