+++ api/src/main/java/me/nexters/liliput/api/infrastructure/InfrastructureConfig.java
@@ -1,0 +1,11 @@
+package me.nexters.liliput.api.infrastructure;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration
+@ComponentScan
+public class InfrastructureConfig {
+}
