+++ api/src/main/java/me/nexters/liliput/api/interfaces/v1/dto/response/V1ShortUrlResponse.java
@@ -1,7 +1,7 @@
 package me.nexters.liliput.api.interfaces.v1.dto.response;
 
 import lombok.Data;
-import me.nexters.liliput.api.domain.shorten.service.ShortUrlModel;
+import me.nexters.liliput.api.domain.dto.ShortUrlModel;
 
 @Data
 public class V1ShortUrlResponse {
