+++ api/src/main/resources/templates/sample-success.ftl
@@ -1,0 +1,15 @@
+<!doctype html>
+<head>
+    <meta charset="utf-8">
+    <title>papao</title>
+</head>
+
+<body>
+<div>로그인 성공</div>
+<div>id : ${id?c}</div>
+<div>tokenRefreshInterval : ${tokenRefreshInterval}</div>
+<div>accessToken : ${accessToken}</div>
+<div>phone : ${profile.phone.nationalNumber}</div>
+<div>email : ${profile.email.address}</div>
+<div>application : ${profile.application.id}</div>
+</body>
